+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2026/03/30
+  * 修复popup的active_window_by_parent_widget比popup先销毁时，会发生崩溃的问题(感谢兆坤提供补丁)
+
 2026/03/27
   * 完善zip_file_t的注释(感谢福明提供补丁)
   * object_locker支持clear_props(感谢雨欣提供补丁)